refactor(gemini): extract the embedding path into embed.go
Move the batchEmbedContents types, geminiModelResource, and Embed into embed.go
(mirroring the existing image.go split), bringing gemini.go back under the
800-line house limit. No behavior change.
